+++ packages/orq-rc/src/orq_ai_sdk/contacts.py
@@ -70,9 +70,6 @@ class Contacts(BaseSDK):
             user_agent_header="user-agent",
             accept_header_value="application/json",
             http_headers=http_headers,
-            _globals=models.CreateContactGlobals(
-                contact_id=self.sdk_configuration.globals.contact_id,
-            ),
             security=self.sdk_configuration.security,
             get_serialized_body=lambda: utils.serialize_request_body(
                 request, False, False, "json", models.CreateContactRequestBody
@@ -186,9 +183,6 @@ class Contacts(BaseSDK):
             user_agent_header="user-agent",
             accept_header_value="application/json",
             http_headers=http_headers,
-            _globals=models.CreateContactGlobals(
-                contact_id=self.sdk_configuration.globals.contact_id,
-            ),
             security=self.sdk_configuration.security,
             get_serialized_body=lambda: utils.serialize_request_body(
                 request, False, False, "json", models.CreateContactRequestBody