+++ packages/orq-rc/src/orq_ai_sdk/feedback.py
@@ -61,9 +61,6 @@ class Feedback(BaseSDK):
             user_agent_header="user-agent",
             accept_header_value="application/json",
             http_headers=http_headers,
-            _globals=models.CreateFeedbackGlobals(
-                contact_id=self.sdk_configuration.globals.contact_id,
-            ),
             security=self.sdk_configuration.security,
             get_serialized_body=lambda: utils.serialize_request_body(
                 request, False, False, "json", models.CreateFeedbackRequestBody
@@ -168,9 +165,6 @@ class Feedback(BaseSDK):
             user_agent_header="user-agent",
             accept_header_value="application/json",
             http_headers=http_headers,
-            _globals=models.CreateFeedbackGlobals(
-                contact_id=self.sdk_configuration.globals.contact_id,
-            ),
             security=self.sdk_configuration.security,
             get_serialized_body=lambda: utils.serialize_request_body(
                 request, False, False, "json", models.CreateFeedbackRequestBody