+++ packages/orq-rc/src/orq_ai_sdk/models/findonebykeypromptsnippetop.py
@@ -10,23 +10,11 @@ from orq_ai_sdk.types import (
     UNSET,
     UNSET_SENTINEL,
 )
-from orq_ai_sdk.utils import FieldMetadata, HeaderMetadata, PathParamMetadata
+from orq_ai_sdk.utils import FieldMetadata, PathParamMetadata
 import pydantic
 from pydantic import model_serializer
 from typing import Any, Dict, List, Literal, Optional, Union
 from typing_extensions import Annotated, NotRequired, TypeAliasType, TypedDict
-
-
-class FindOneByKeyPromptSnippetGlobalsTypedDict(TypedDict):
-    contact_id: NotRequired[str]
-
-
-class FindOneByKeyPromptSnippetGlobals(BaseModel):
-    contact_id: Annotated[
-        Optional[str],
-        pydantic.Field(alias="contactId"),
-        FieldMetadata(header=HeaderMetadata(style="simple", explode=False)),
-    ] = None
 
 
 class FindOneByKeyPromptSnippetRequestTypedDict(TypedDict):
@@ -1319,7 +1307,7 @@ class FindOneByKeyPromptSnippetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2025-02-10T17:12:36.462Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2025-02-10T22:25:09.383Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")